pyproject.toml
[project]
name = "extraction-engine"
version = "0.1.0"
description = "Eval-first document extraction engine using OCR + VLM two-pass approach"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    # Anthropic
    "anthropic>=0.44.0",
    # OCR (Tesseract - default, free, local)
    "pytesseract>=0.3.10",
    # Database
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.14.0",
    # Task queue
    "celery[redis]>=5.4.0",
    # Document processing
    "Pillow>=11.0.0",
    "pymupdf>=1.25.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.19",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "python-Levenshtein>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]
eval = [
    "deepeval>=2.0.0",
]
azure = [
    "azure-ai-documentintelligence>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extraction_engine"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
